Penambahan Home dan Landing Navbar
