+++ package.json
@@ -10,6 +10,7 @@
     "bootstrap": "^5.1.3",
     "cors": "^2.8.5",
     "react": "^18.2.0",
+    "react-bootstrap": "^2.4.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
