+++ src/pages/Home.js
@@ -1,9 +1,9 @@
-import HeaderNav from "../components/navbar/HeaderNav";
+import LandingNav from "../components/navbar/HomeNav";
 
 function Home() {
   return (
     <>
-      < HeaderNav />
+      < LandingNav />
       <h1>HOME PAGE</h1>
       <a href="/rock-paper-scissors">Rock Paper Scissor</a>
     </>
