--- src/pages/Profile.js
@@ -1,9 +1,0 @@
-const Profile = () => {
-    return (
-        <div>
-            <h1>Profile Page</h1>
-        </div>
-    )
-}
-
-export default Profile
